--- src/common/Banuba/ClientToken.ts
@@ -1,6 +1,0 @@
-declare global {
-  interface Window { BANUBA_CLIENT_TOKEN: string }
-}
-window.BANUBA_CLIENT_TOKEN = process.env.REACT_APP_BANUBA_KEY as string;
-
-export {};
