+++ src/common/notify.ts
@@ -1,0 +1,13 @@
+import { toast } from 'react-toastify';
+
+export const notifyError = (message: string): void => {
+  toast.error(message, {
+    position: 'top-right',
+    autoClose: 5000,
+    hideProgressBar: false,
+    closeOnClick: true,
+    pauseOnHover: true,
+    draggable: true,
+    progress: undefined,
+  });
+};
