+++ src/common/providers/Interfaces.ts
@@ -6,18 +6,3 @@ export interface ICall{
   isReceivingCall: boolean,
   name: string,
 }
-
-export type ContextType = {
-  call: ICall | undefined,
-  callAccepted: boolean,
-  userVideo: React.MutableRefObject<{srcObject: MediaStream} | React.LegacyRef<HTMLVideoElement> | undefined>,
-  stream: MediaStream | undefined,
-  setStream: React.Dispatch<React.SetStateAction<MediaStream | undefined>>,
-  name: string,
-  setName: React.Dispatch<React.SetStateAction<string>>,
-  callEnded: boolean,
-  me: string,
-  callUser: (id: string) => void,
-  leaveCall: () => void,
-  answerCall: () => void,
-}
